refactor(qat): move BatchNormedActivatedConv1d from layers module to blocks module
